reformat(CoinView): view 클래스 분리
- 동전의 개수를 출력하는 클래스를 CoinView로 이동
