+++ src/main/java/vendingmachine/controller/ChangeController.java
@@ -1,23 +1,16 @@
 package vendingmachine.controller;
 
-import java.util.Map;
-
 import vendingmachine.repository.Change;
-import vendingmachine.domain.Coin;
 import vendingmachine.service.ChangeService;
-import vendingmachine.view.VendingMachineAccountView;
+import vendingmachine.view.CoinView;
 
 public class ChangeController {
-	private static final String CHANGE_MESSAGE = "잔돈";
-
 	public static void setChange() {
-
 		ChangeService.calculateChange();
 	}
 
 	public static void printChange() {
-		System.out.println(CHANGE_MESSAGE);
-		Map<Coin, Integer> change = Change.getChange();
-		VendingMachineAccountView.printCoinMap(change);
+		CoinView.printChange();
+		CoinView.printCoinMap(Change.getChange());
 	}
 }
