+++ src/main/java/vendingmachine/view/VendingMachineAccountView.java
@@ -1,13 +1,8 @@
 package vendingmachine.view;
 
-import java.util.Map;
-
-import vendingmachine.domain.Coin;
-
 public class VendingMachineAccountView {
 	private static final String INPUT_MESSAGE = "자판기가 보유하고 있는 금액을 입력해 주세요.";
 	private static final String VENDING_MACHINE_COIN_MESSAGE = "자판기가 보유한 동전";
-	private static final String COIN_LIST_FORMAT = "%d원 - %d개";
 
 	public static void printInputGuide() {
 		System.out.println(INPUT_MESSAGE);
@@ -17,12 +12,4 @@ public class VendingMachineAccountView {
 		System.out.println(VENDING_MACHINE_COIN_MESSAGE);
 	}
 
-	public static void printCoinMap(Map<Coin, Integer> coinMap) {
-		for (Coin coin : Coin.values()) {
-			if (coinMap.containsKey(coin)) {
-				System.out.println(String.format(COIN_LIST_FORMAT, coin.getAmount(), coinMap.get(coin)));
-			}
-		}
-		System.out.println();
-	}
 }
